pyproject.toml
[project]
name = "rag-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-experimental>=0.3.4",
    "langchain-neo4j>=0.4.0",
    "langchain-openai>=0.3.16",
    "mem0ai>=0.1.97",
    "neo4j>=5.28.1",
    "openai>=1.77.0",
    "rank-bm25>=0.2.2",
]
